+++ models/grupoetapa.js
@@ -1,8 +1,11 @@
 'use strict';
+
+const errors = require('../utils/errors');
+
 module.exports = function(sequelize, DataTypes) {
-  var GrupoEtapa = sequelize.define('GrupoEtapa', {
+  let GrupoEtapa = sequelize.define('GrupoEtapa', {
     fechaInicio : {
-      type : DataTypes.DATE,
+      type      : DataTypes.DATE,
       allowNull : true
     },
     fechaFin : {
@@ -44,22 +47,32 @@ module.exports = function(sequelize, DataTypes) {
       ////////////////////////////////////
       crearGrupoEtapaT: function(idGrupo, idEtapa, transaction){
         return new Promise( (resolve, reject) => {
+          if ( !idGrupo )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id del grupo') );
+          if ( idGrupo < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id del grupo inválido') );
+          if ( !idEtapa )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id de la etapa') );
+          if ( idEtapa < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id de etapa inválido') );
           return this.create({
-            EtapaId: idEtapa,
-            GrupoId: idGrupo,
+            EtapaId    : idEtapa,
+            GrupoId    : idGrupo,
             fechaInicio: new Date(),
-            fechaFin: null
+            fechaFin   : null
           }, { transaction : transaction})
           .then( (registro) => {
             return resolve(registro);
           })
-          .catch( error => {
-            return reject(error);
+          .catch( fail => {
+            return reject( errors.ERROR_HANDLER(fail) );
           });
         });
       },
       cambiarGrupoDeEtapaT: function(idGrupo, idEtapaAntigua, idEtapaNueva, transaction){
         return new Promise( (resolve, reject) => {
+          if ( !idGrupo )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id del grupo') );
+          if ( idGrupo < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id del grupo inválido') );
+          if ( !idEtapaAntigua )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id de la etapa antigua') );
+          if ( idEtapaAntigua < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id de etapa antigua inválido') );
+          if ( !idEtapaNueva )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id de la etapa nueva') );
+          if ( idEtapaNueva < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id de etapa nueva inválido') );
           return this.update({
             fechaFin : new Date()
           }, {
@@ -71,25 +84,27 @@ module.exports = function(sequelize, DataTypes) {
           })
           .then( registro => {
             return this.create({
-              GrupoId: idGrupo,
-              EtapaId: idEtapaNueva,
-              fechaInicio: new Date(),
-              fechaFin: null
+              GrupoId     : idGrupo,
+              EtapaId     : idEtapaNueva,
+              fechaInicio : new Date(),
+              fechaFin    : null
             }, { transaction : transaction })
             .then( registro2 => {
               return resolve(registro2);
             })
-            .catch( error2 => {
-              return reject(error2);
+            .catch( fail2 => {
+              return reject( errors.ERROR_HANDLER(fail2) );
             });
           })
-          .catch( error => {
-            return reject(error);
+          .catch( fail => {
+            return reject( errors.ERROR_HANDLER(fail) );
           });
         });
       },
       eliminarRegistrosDeGrupoT: function(idGrupo, transaction){
         return new Promise( (resolve, reject) => {
+          if ( !idGrupo )    return reject( errors.SEQUELIZE_FK_ERROR('No ingresó el id del grupo') );
+          if ( idGrupo < 0 ) return reject( errors.SEQUELIZE_FK_ERROR('Id del grupo inválido') );
           return this.destroy({
             where : {
               GrupoId : idGrupo
@@ -97,10 +112,11 @@ module.exports = function(sequelize, DataTypes) {
             transaction : transaction
           })
           .then( resultado => {
-            return resolve(resultado);
+            if( resultado === 0 ) return reject( errors.SEQUELIZE_ERROR('Delete error', 'No se encontró el registro de la etapa del grupo para eliminar') );
+            if( resultado === 1 ) return resolve(resultado);
           })
-          .catch( error => {
-            return reject(error);
+          .catch( fail => {
+            return reject( errors.ERROR_HANDLER(fail) );
           });
         });
       }
